moved TimeUtilsService.java to new package utils, changes by refactoring
