+++ backend/src/main/java/de/neuefische/backend/reward/RewardService.java
@@ -1,5 +1,6 @@
 package de.neuefische.backend.reward;
 
+import de.neuefische.backend.utlis.TimeUtilsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
