+++ backend/src/test/java/de/neuefische/backend/reward/RewardServiceTest.java
@@ -1,5 +1,6 @@
 package de.neuefische.backend.reward;
 
+import de.neuefische.backend.utlis.TimeUtilsService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
@@ -8,7 +9,6 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 
-import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.util.Arrays;
 import java.util.List;
